requirements.txt
python-dotenv==1.0.0
requests==2.31.0
pinecone-client==3.0.2
anthropic==0.7.0
beautifulsoup4==4.12.2
fastapi==0.104.1
uvicorn==0.24.0
numpy<2
pydantic==2.5.0
sentence-transformers==3.0.1
